requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-cache2==0.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
pendulum==3.0.0
pyasn1==0.5.1
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.24
sqlmodel==0.0.14
starlette==0.32.0.post1
time-machine==2.13.0
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.25.0
